CO-4457
refactor: merged in changes on DomainConfig from the origin/development branch

feature: seemingly working as a SpringBoot/Vaadin project

refactor: cosmetic refactoring :)

refactor: changed delimiter to new line for domainConfig notes

testing

Added the Dockerfile and the run.sh file. Adjusted the .pom and and Dockerfile to the project.

Added the Dockerfile and the run.sh file. Adjusted the .pom and and Dockerfile to the project.

fix: closer to functioning, deployable Docker image

fix: oops forgot this one :(

fix: a working docker container :)